+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## lifebit-ai/cloudos-cli: changelog
 
+## v2.11.0 (2024-04-16)
+
+- Now, the default `cloudos job run` command will save job process logs. To prevent saving process logs, you can use the new flag `--do-not-save-logs`.
+- Removes unsupported `--spot` option from `cloudos job run`.
+
 ## v2.10.0 (2024-04-11)
 
 - Adds the new parameter `--workflow-docs-link` to add a documentation link to the imported workflow.